Add a demo of an ANCF shell element toroidal wheel model.
It contains a tread made up of shell elements constrained to the rim through position and direction constraints. The direction in the implementation of the ANCF shell element must be changed to the Y axis for the model to work properly. The wheel is pushed solely by initial velocities and the contact is node-to-surface including simple Coulomb friction.